chore(delete): delete a useless .hpp file, and fix thread functions
