+++ CAPI/Communication/Communication.vcxproj
@@ -20,7 +20,6 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="include\Communication.h" />
-    <ClInclude Include="include\message.hpp" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\Communication.cpp" />
