+++ CAPI/Communication/Communication.vcxproj.filters
@@ -18,9 +18,6 @@
     <ClInclude Include="include\Communication.h">
       <Filter>头文件</Filter>
     </ClInclude>
-    <ClInclude Include="include\message.hpp">
-      <Filter>头文件</Filter>
-    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\Communication.cpp">
