+++ CAPI/Communication/include/Communication.h
@@ -28,6 +28,10 @@
 // 3.ClientCommunication依赖于MultiThreadClientCommunication的接口，而MultiThreadClientCommunication依赖于ClientCommunication的实体
 // 在THUAI4中，从MultiThreadClientCommunication(THUAI4中名为Communication)构造ClientCommunication(THUAI4中名为CAPI)中，又用到了大量的回调函数
 // 所以不妨将ClientCommunication依赖于MultiThreadClientCommunication的部分抽取为一个接口ICommunication
+using pointer_m2c = std::variant<std::shared_ptr<Protobuf::MessageToClient>, std::shared_ptr<Protobuf::MessageToOneClient>, std::shared_ptr<Protobuf::MessageToInitialize>, std::nullptr_t>; // 类型安全的联合体
+#define TYPEM2C 0
+#define TYPEM2OC 1
+#define TYPEM2I 2
 
 /// <summary>
 /// 信息的委托接口，是指ClientCommunication依赖于MultiThreadClientCommunication的部分
@@ -156,7 +160,7 @@ public:
     ~MultiThreadClientCommunication() {};
 
     /// <summary>
-    /// 在MultiThreadClientCommunication对象构造完毕后设置std::unique_ptr<ClientCommunication> capi的内容
+    /// 在MultiThreadClientCommunication对象构造完毕后设置std::unique_ptr/<ClientCommunication/> capi的内容
     /// </summary>
     void init();
